guestbook/09_entitygroup/guestbook_entitygroup_01/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.9.0